chore: seperate event dsl
