+++ src/main/kotlin/dev/fruxz/sparkle/framework/event/dsl/SparkleEventDSL.kt
@@ -1,0 +1,5 @@
+package dev.fruxz.sparkle.framework.event.dsl
+
+@DslMarker
+@MustBeDocumented
+annotation class SparkleEventDSL
